+++ glompo/benchmark_fncs/trigonometric02.py
@@ -8,7 +8,7 @@ from ._base import BaseTestCase
 class Trigonometric(BaseTestCase):
     """ When called returns evaluations of the Trigonometric02 function. """
 
-    def __init__(self, dims: int = 2, delay: int = 0):
+    def __init__(self, dims: int = 2, delay: float = 0):
         """
         Implementation of the Trigonometric02 optimization test function.
         Recommended bounds: [-500, 500] * dims
@@ -20,7 +20,7 @@ class Trigonometric(BaseTestCase):
         ----------
         dims: int = 2
             Number of dimensions of the problem
-        delay: int = 0
+        delay: float = 0
             Delay in seconds after the function is called before results are returned.
             Useful to simulate harder problems.
         """
